requirements.txt
absl-py==0.9.0
appdirs==1.4.4
appnope==0.1.0
arrow==0.15.6
asgiref==3.2.10
astunparse==1.6.3
attrs==19.3.0
autopep8==1.5.3
backcall==0.1.0
beautifulsoup4==4.9.0
binaryornot==0.4.4
blis==0.4.1
boto3==1.14.26
botocore==1.17.26
cached-property==1.5.1
cachetools==4.1.0
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cookiecutter==1.7.2
cycler==0.10.0
cymem==2.0.3
Cython==0.29.14
decorator==4.4.1
defusedxml==0.6.0
Django==3.0.8
django-filter==2.3.0
django-markdown2==0.3.1
django-polls @ file:///C:/DjangoProject/django-polls/dist/django-polls-0.1.tar.gz
django-storages==1.9.1
docutils==0.15.2
edx-dl==0.1.13
Flask==1.1.2
gast==0.3.3
get==2019.4.13
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
gunicorn==20.0.4
h5py==2.10.0
html5lib==1.0.1
idna==2.9
importlib-metadata==1.4.0
ipdb==0.12.3
ipython==7.11.1
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.11.2
jinja2-time==0.2.0
jmespath==0.10.0
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.5.2
Markdown==3.2.1
markdown2==2.3.9
MarkupSafe==1.1.1
matplotlib==3.1.2
more-itertools==8.1.0
murmurhash==1.0.2
mysql-connector-python==8.0.20
numpy==1.18.3
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==0.25.3
parso==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plac==1.1.3
post==2019.4.13
poyo==0.5.0
preshed==3.0.2
prompt-toolkit==3.0.2
protobuf==3.11.3
psycopg2==2.8.5
ptyprocess==0.6.0
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
Pygments==2.5.2
PyMySQL==0.9.3
pyodbc==4.0.30
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.15.0
python-slugify==4.0.0
pytz==2020.1
PyYAML==5.3.1
query-string==2019.4.13
reportlab==3.5.42
request==2019.4.13
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.0
s3transfer==0.3.3
scipy==1.4.1
seaborn==0.10.0
sentencepiece==0.1.85
six==1.15.0
soupsieve==2.0
spacy==2.2.3
sqlparse==0.3.1
srsly==1.0.1
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0rc4
tensorflow-estimator==2.2.0
termcolor==1.1.0
text-unidecode==1.3
thinc==7.3.1
toml==0.10.1
torch @ https://download.pytorch.org/whl/cpu/torch-1.4.0%2Bcpu-cp38-cp38-win_amd64.whl
torchtext==0.5.0
torchvision==0.5.0
tqdm==4.41.1
traitlets==4.3.3
urllib3==1.25.10
vcstools==0.1.42
virtualenv==16.7.8
virtualenvwrapper-win==1.2.5
wasabi==0.6.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
wstool==0.1.17
youtube-dl==2020.5.8
zeep==3.4.0
zipp==2.0.1
